3.1_rag/requirements.txt
sentence-transformers
faiss-cpu
transformers
torch
beautifulsoup4
requests
tqdm
